Store the last visited UI in a cookie. This cookies value (when present) is used for the redirect when the root URL is requested. This allows the preferred UI top be remembered.
The cookie is set server-side on all UI requests. The standard and alternate UIs also set the cookie from javascript to defeat caching - if /s/rua/ and /a/rua/ are both cached then switching between them won't cause the server to set the cookie - in this case the javascript code will do it anyway. As the basic UI doesn't use JavaScript at all it relies on the server setting the cookie. Its only really an issue if someone switches to a UI they've not been to before and then switches back (in which case the switch back won't be stored in the cookie unless they navigate elsewhere in the UI to a page that isn't cached).