Change the output target to zxweather (from desktop).
Also changed config file loading - on windows it will now look for a file with
a matching basename to the executable in the executables directory by default.
If a config file with the old name (zxweather-desktop.ini in the current
directory) exists it will be renamed.

On unix, ~/.zxweather-desktop.ini will be loaded instead.