Initial support for extra sensors in the desktop app. App builds and runs and the extra sensors work in charts and view data though the UI isn't great.
Currently the sensors are always visible regardless of station type or
configuration. They're not present in the create chart window or export data
window.

The implementation is a bit messy due to running out of space in the
SampleColumns enum. As 64bit enums aren't an option right now this has meant
adding a second ExtraColumns enum, wrapping up both in a struct and in many
cases duplicating code to deal with the two enums separately. This will all be
cleaned up in the future when we can either make some sort of 64bit QFlags work
or change the way columns (and probably sample sets) are handled entirely.

This changeset also removes the CustomiseChart dialog - its not been used in
years, possibly doesn't work anymore (was disabled as part of the multi-dataset
work), isn't finished and probably never will be.