desktop: WebDataSource now tries not to bother the web server when asked for samples.
It now scans for gaps in downloaded data files and if there are none they'll be considered 'complete' and will never be downloaded from the server again (at least until the cache database is cleared).

When checking query timespan the cache database will be consulted first and if the timespan fits within the cache database it won't bother asking the web server for its timespan.

It will also only load sysconfig once per session.

All of this means that a chart for last month can be plotted without a single web request being made. This all speeds up charts, export, view data and reports quite a bit in some cases.