desktop: 'Fix' bug where sometimes a video refuses to play (just stops immediately after you click play).
This bug isn't really fixed - its still there and I've still no idea what causes it. There just seems to be a random chance of a new QMediaPlayer simply being broken in this way. Its detectable by it raising an EndOfMedia status without actually being at the end of the media. The easiest solution is to just delete and recreate the QMediaPlayer and try again when this happens.

When the VideoPlayer widget encounters an InvalidMedia status it will now try to reload the media up to two seconds with a one second delay each time. This may be enough to make the other unrelated 'Invalid Media' problem go away too (assuming its just caused by the video not being fully written to disk when the QMediaPlayer tries to read it).