Desktop: WebCacheDB: Don't bother server if query is covered by CacheDB.
Add some new columns to store how much contiguous data there is add the start and end of each data file along with the timespan the data file should cover if it is complete.

Added a new function to WebCacheDB to query this information and check that all samples within a timespan are covered by sufficiently cached data files.

Updated WebDataSource to check if timespans are cached before talking to the web server.

This fixes #7