Database, Desktop: Function, Table and Report for gaps in the sample table
Added a new table for recoding the existence of gaps in the sample table and optionally giving them a label/description.

Added SQL function for fetching all gaps in the sample table for a particular station (whether they're in the gaps table or not)

Added a postgres-only report for the desktop client which calls the SQL function and shows all gaps for the station.

This is the first part of #117.