Web UI: Support archived stations
The station overview page will now show the station archived message (if there is one) and archived date instead of a generic station offline message. You can replace the archived page with a custom one by creating an "archived.html" file in the directory for the archived station. An example file is provided.