DavisLogger: Support Rev. A firmware. In theory.
I've got no access to hardware so I can't confirm it works. I also can't easily write unit tests as I've got no samples of data it work with - all I could test is my implementation matches my understanding of the spec. So Rev. A firmware support is 100% untested and as such it emits a warning if this firmware is detected.