DavisLogger: Now receives LOOP2 packets (#131). Also adds sea level pressure columns (#140)
And makes that data available via new columns in the database as well as RabbitMQ.

This commit also adds database columns for mean sea level pressure. The appropriate values are now logged to the appropriate places for live data.

Davis logger for now logs the same value to both absolute and mean sea level pressure columns for samples rather than leaving the absolute pressure column as null. This will be changed near the end of ticket #140 before the data migration step.