Server: Support for mean sea level pressure and LOOP2 data
The new mean sea level and LOOP2 fields are opt-in. The subscribe command now has optional version parameters for the /live and /samples qualifiers. The default version is 1 which does not include the extra fields which means compatibility with older versions of the desktop client and Web UI are unaffected.

As part of this production of the CSV records has been moved to functions in the database.